[Update] Explore Category Item Sorting - Fixed an issue with ExploreCategory sorting - Made ExploreCategoryItem sortable

+++ database/migrations/2019_06_25_195950_create_explore_categories_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->string('slug');
             $table->string('type');
             $table->string('size');
-            $table->smallInteger('position');
+            $table->integer('position');
             $table->boolean('is_global');
             $table->boolean('is_enabled');
             $table->timestamps();
